refactor: separate entity packages.
