+++ dialogy/types/entity/base_entity/__init__.py
