+++ dialogy/types/entity/deserialize/__init__.py
@@ -1,13 +1,13 @@
 from typing import Any, Dict, Optional, Union
 
 from dialogy import constants as const
-from dialogy.types.entity.credit_card_number_entity import CreditCardNumberEntity
-from dialogy.types.entity.currency_entity import CurrencyEntity
-from dialogy.types.entity.duration_entity import DurationEntity
-from dialogy.types.entity.numerical_entity import NumericalEntity
-from dialogy.types.entity.people_entity import PeopleEntity
-from dialogy.types.entity.time_entity import TimeEntity
-from dialogy.types.entity.time_interval_entity import TimeIntervalEntity
+from dialogy.types.entity.credit_card_number import CreditCardNumberEntity
+from dialogy.types.entity.amount_of_money import CurrencyEntity
+from dialogy.types.entity.duration import DurationEntity
+from dialogy.types.entity.numerical import NumericalEntity
+from dialogy.types.entity.people import PeopleEntity
+from dialogy.types.entity.time import TimeEntity
+from dialogy.types.entity.time_interval import TimeIntervalEntity
 
 
 def deserialize_duckling_entity(
