+++ dialogy/types/entity/duration/__init__.py
