+++ dialogy/types/entity/numerical/__init__.py
