+++ dialogy/types/entity/time/__init__.py
@@ -14,7 +14,7 @@ import attr
 import pydash as py_
 
 from dialogy import constants as const
-from dialogy.types.entity.numerical_entity import NumericalEntity
+from dialogy.types.entity.numerical import NumericalEntity
 
 
 @attr.s
