+++ dialogy/types/entity/time_interval/__init__.py
